[wip] QA step 3 to create elb2 passed
